VThreadID: change to uintptr_t

+++ open-vm-tools/lib/include/vthreadBase.h
@@ -82,7 +82,7 @@ extern "C" {
  * Types
  */
 
-typedef uint64 VThreadID;
+typedef uintptr_t VThreadID;
 
 #define VTHREAD_INVALID_ID    (VThreadID)(0)
